pyproject.toml
[tool.poetry]
name = "chatpayment"
version = "1"
description = "click and payments connection chat"
authors = ["Azamat"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
